test-mcps/s3-mcp/pyproject.toml
[project]
name = "s3-mcp"
version = "0.1.0"
description = "AWS S3 MCP server for LangSight testing"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
